Entities can now be enabled/disabled. Don't bind vertex/index buffer if it's already bound.

+++ Engine/Graphics/src/Graphics/RHI/DX12/RHI_Pipeline_DX12.cpp
@@ -61,7 +61,7 @@ namespace Insight
 					PolygonModeToDX12(pso.PolygonMode),
 					CullModeToDX12(pso.CullMode),
 					pso.FrontFace == FrontFace::CounterClockwise ? true : false,
-					pso.DepthConstantBaisValue,
+					static_cast<INT>(pso.DepthConstantBaisValue),
 					D3D12_DEFAULT_DEPTH_BIAS_CLAMP,
 					pso.DepthSlopeBaisValue,
 					TRUE,
